backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
msgpack==1.1.1
multidict==6.6.3
nest-asyncio==1.6.0
packaging==25.0
Pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pygments==2.19.2
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-django==4.11.1
pytest-mock==3.14.1
python-decouple==3.8
python-telegram-bot==20.7
qrcode==8.2
redis==6.2.0
requests==2.32.4
requests-mock==1.12.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
twisted==25.5.0
txaio==25.6.1
typing-extensions==4.14.0
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zope-interface==7.2

